apps/api/pyproject.toml
[project]
name = "mealplanner-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  # FIX #1 (review Phase 1 2026-04-12) : dépendance vers le package DB partagé
  # NOTA devops-engineer : ajouter packages/db dans [tool.uv.workspace] members du pyproject.toml racine
  "mealplanner-db",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "pgvector>=0.3",
  "supabase>=2.6",
  "python-jose[cryptography]>=3.3",
  "slowapi>=0.1.9",
  "redis>=5.0",
  "loguru>=0.7",
  "httpx>=0.27",
  "tenacity>=8.3",
  "orjson>=3.10",
  # Phase 2 — Billing Stripe
  "stripe>=7.0",
  # Phase 2 — Stockage objet MinIO/R2 (presigned URLs, upload PDFs)
  "boto3>=1.34",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=4.1",
  "httpx>=0.27",
  "faker>=25",
]

[tool.uv.sources]
mealplanner-db = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
